+++ lib/context/context.dart
@@ -6,6 +6,7 @@ import '../version_control.dart' show sdkPackage;
 import './context_app.dart' show contextApp;
 import './context_device.dart' show contextDevice;
 import './context_locale.dart' show contextLocale;
+import './context_location.dart' show ContextLocation;
 import './context_network.dart' show contextNetwork;
 import './context_os.dart' show contextOS;
 
@@ -17,6 +18,7 @@ class Context {
   Future<Map<String, dynamic>> toMap() async => {
         ...await _base,
         'locale': await contextLocale(),
+        'location': ContextLocation().toJson(),
         'network': await contextNetwork()
       };