+++ lib/lifecycle/lifecycle.dart
@@ -10,25 +10,29 @@ class AppLifecycle {
   factory AppLifecycle() => _appLifecycle;
 
   AppLifecycle._internal()
-      : _callbacks = [],
+      : _subscriptions = [],
         _state = AppLifecycleState.resumed;
-
-  /// @nodoc
-  AppLifecycleState get state => _state;
-  set state(AppLifecycleState state) {
-    _state = state;
-
-    for (final cb in _callbacks) {
-      cb(state);
-    }
-  }
 
   static final _appLifecycle = AppLifecycle._internal();
 
-  final List<void Function(AppLifecycleState state)> _callbacks;
+  final List<void Function(AppLifecycleState state)> _subscriptions;
   AppLifecycleState _state;
 
   /// @nodoc
-  void addCallback(void Function(AppLifecycleState state) onLifecycleState) =>
-      _callbacks.add(onLifecycleState);
+  AppLifecycleState get state => _state;
+  set state(AppLifecycleState newState) {
+    if (newState == null || newState == _state) {
+      return;
+    }
+
+    _state = newState;
+
+    for (final subscription in _subscriptions) {
+      subscription(newState);
+    }
+  }
+
+  /// @nodoc
+  void subscribe(void Function(AppLifecycleState state) onAppLifecycleState) =>
+      _subscriptions.add(onAppLifecycleState);
 }